requirements.txt
streamlit
openai
pydantic
requests
python-dotenv
pyparsing